+++ examples/test_examples.py
@@ -93,34 +93,6 @@ class ExamplesTestsAccelerate(unittest.TestCase):
             self.assertTrue(os.path.isfile(os.path.join(tmpdir, "unet", "diffusion_pytorch_model.safetensors")))
             self.assertTrue(os.path.isfile(os.path.join(tmpdir, "scheduler", "scheduler_config.json")))
 
-    def test_textual_inversion(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            test_args = f"""
-                examples/textual_inversion/textual_inversion.py
-                --pretrained_model_name_or_path hf-internal-testing/tiny-stable-diffusion-pipe
-                --train_data_dir docs/source/en/imgs
-                --learnable_property object
-                --placeholder_token <cat-toy>
-                --initializer_token a
-                --validation_prompt <cat-toy>
-                --validation_steps 1
-                --save_steps 1
-                --num_vectors 2
-                --resolution 64
-                --train_batch_size 1
-                --gradient_accumulation_steps 1
-                --max_train_steps 2
-                --learning_rate 5.0e-04
-                --scale_lr
-                --lr_scheduler constant
-                --lr_warmup_steps 0
-                --output_dir {tmpdir}
-                """.split()
-
-            run_command(self._launch_args + test_args)
-            # save_pretrained smoke test
-            self.assertTrue(os.path.isfile(os.path.join(tmpdir, "learned_embeds.safetensors")))
-
     def test_unconditional_checkpointing_checkpoints_total_limit(self):
         with tempfile.TemporaryDirectory() as tmpdir:
             initial_run_args = f"""
@@ -196,104 +168,4 @@ class ExamplesTestsAccelerate(unittest.TestCase):
             self.assertEqual(
                 {x for x in os.listdir(tmpdir) if "checkpoint" in x},
                 {"checkpoint-8", "checkpoint-10", "checkpoint-12"},
-            )
-
-    def test_textual_inversion_checkpointing(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            test_args = f"""
-                examples/textual_inversion/textual_inversion.py
-                --pretrained_model_name_or_path hf-internal-testing/tiny-stable-diffusion-pipe
-                --train_data_dir docs/source/en/imgs
-                --learnable_property object
-                --placeholder_token <cat-toy>
-                --initializer_token a
-                --validation_prompt <cat-toy>
-                --validation_steps 1
-                --save_steps 1
-                --num_vectors 2
-                --resolution 64
-                --train_batch_size 1
-                --gradient_accumulation_steps 1
-                --max_train_steps 3
-                --learning_rate 5.0e-04
-                --scale_lr
-                --lr_scheduler constant
-                --lr_warmup_steps 0
-                --output_dir {tmpdir}
-                --checkpointing_steps=1
-                --checkpoints_total_limit=2
-                """.split()
-
-            run_command(self._launch_args + test_args)
-
-            # check checkpoint directories exist
-            self.assertEqual(
-                {x for x in os.listdir(tmpdir) if "checkpoint" in x},
-                {"checkpoint-2", "checkpoint-3"},
-            )
-
-    def test_textual_inversion_checkpointing_checkpoints_total_limit_removes_multiple_checkpoints(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            test_args = f"""
-                examples/textual_inversion/textual_inversion.py
-                --pretrained_model_name_or_path hf-internal-testing/tiny-stable-diffusion-pipe
-                --train_data_dir docs/source/en/imgs
-                --learnable_property object
-                --placeholder_token <cat-toy>
-                --initializer_token a
-                --validation_prompt <cat-toy>
-                --validation_steps 1
-                --save_steps 1
-                --num_vectors 2
-                --resolution 64
-                --train_batch_size 1
-                --gradient_accumulation_steps 1
-                --max_train_steps 3
-                --learning_rate 5.0e-04
-                --scale_lr
-                --lr_scheduler constant
-                --lr_warmup_steps 0
-                --output_dir {tmpdir}
-                --checkpointing_steps=1
-                """.split()
-
-            run_command(self._launch_args + test_args)
-
-            # check checkpoint directories exist
-            self.assertEqual(
-                {x for x in os.listdir(tmpdir) if "checkpoint" in x},
-                {"checkpoint-1", "checkpoint-2", "checkpoint-3"},
-            )
-
-            resume_run_args = f"""
-                examples/textual_inversion/textual_inversion.py
-                --pretrained_model_name_or_path hf-internal-testing/tiny-stable-diffusion-pipe
-                --train_data_dir docs/source/en/imgs
-                --learnable_property object
-                --placeholder_token <cat-toy>
-                --initializer_token a
-                --validation_prompt <cat-toy>
-                --validation_steps 1
-                --save_steps 1
-                --num_vectors 2
-                --resolution 64
-                --train_batch_size 1
-                --gradient_accumulation_steps 1
-                --max_train_steps 4
-                --learning_rate 5.0e-04
-                --scale_lr
-                --lr_scheduler constant
-                --lr_warmup_steps 0
-                --output_dir {tmpdir}
-                --checkpointing_steps=1
-                --resume_from_checkpoint=checkpoint-3
-                --checkpoints_total_limit=2
-                """.split()
-
-            run_command(self._launch_args + resume_run_args)
-
-            # check checkpoint directories exist
-            self.assertEqual(
-                {x for x in os.listdir(tmpdir) if "checkpoint" in x},
-                {"checkpoint-3", "checkpoint-4"},
             )
